** rename service folder pt2 **
