+++ src/KRFApp.test.tsx
@@ -4,6 +4,14 @@ import KRFApp from './components/main/KRFApp';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<KRFApp />, div);
+  ReactDOM.render(<KRFApp Routes = {
+    {
+      Routes: {
+        Home: {
+           Component: () => <></>
+        }
+      }
+    }
+  }/>, div);
   ReactDOM.unmountComponentAtNode(div);
 });
